Stash the changes for configurable number of Presto splits per tablet in a separate branch.
Need to rethink this... while the negative infinity to first split is do-able (no bytes to first split), the opposite end of last split to infinite bytes is not so trivial.  Maybe we can get at the last key in the split from some other API instead of the default?